Original SPSS scripts used by Imagen
SPSS files as sent in a ZIP file on 14 February 2017.

These files were probably used only for the pilot phase of Imagen
before 2010.

Then the initial database team swiftly changed to SQL code applied
to an intermediate PostgreSQL database, which has since disappeared.
BL and FU1 Psytools data available from the old Imagen XNAT server
were probably processed using the SQL code.

For FU2 and the previous releases of the new Imagen v2 server,
we switched to an unpublished Python script which is far from
perfect.

For FU3 and upcoming releases of the Imagen v2 server we want to
re-use the R scripts written by John Rogers for the c-VEDA project.

As you can see, these SPSS files are mainly of historical interest.
They're maintained here as a mere reference.